requirements.txt
flask==3.0.3
flask-cors==4.0.1
ultralytics==8.2.48
deepface==0.0.92
pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
pyodbc==5.1.0